Inline lwheader.lua into lwaftr.lua
